Changed app.py to routes.py. routes.py will have routes only
