+++ routes.py
@@ -1,0 +1,10 @@
+from flask import Flask, render_template
+app = Flask(__name__)
+
+@app.route('/')
+def index():
+    return render_template('index.html')
+
+if __name__ == '__main__':
+    # app.secret_key='secret123'
+    app.run(debug = True)
